dumpert/requirements.txt
django
django-filer
django-csvimport